+++ composer.json
@@ -18,12 +18,12 @@
     ],
     "autoload": {
         "psr-4": {
-            "Pear\\Calendar\\": "Calendar/"
+            "PEAR\\Calendar\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Pear\\Calendar\\Test\\": "tests/"
+            "PEAR\\Calendar\\Test\\": "tests/"
         }
     },
     "description": "More info available on: http://pear.php.net/package/Calendar",
