+++ src/Calendar.php
@@ -36,7 +36,7 @@
  * @version   CVS: $Id$
  * @link      http://pear.php.net/package/Calendar
  */
-namespace Pear\Calendar;
+namespace PEAR\Calendar;
 
 /**
  * Allows Calendar include path to be redefined
