+++ src/Decorator/Wrapper.php
@@ -1,0 +1,104 @@
+<?php
+/* vim: set expandtab tabstop=4 shiftwidth=4: */
+
+/**
+ * Contains the Calendar_Decorator_Wrapper class
+ *
+ * PHP versions 4 and 5
+ *
+ * LICENSE: Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. The name of the author may not be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR "AS IS" AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE FREEBSD PROJECT OR CONTRIBUTORS BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Date and Time
+ * @package   Calendar
+ * @author    Harry Fuecks <hfuecks@phppatterns.com>
+ * @author    Lorenzo Alberton <l.alberton@quipo.it>
+ * @copyright 2003-2007 Harry Fuecks, Lorenzo Alberton
+ * @license   http://www.debian.org/misc/bsd.license  BSD License (3 Clause)
+ * @version   CVS: $Id$
+ * @link      http://pear.php.net/package/Calendar
+ */
+namespace PEAR\Calendar\Decorator;
+
+use PEAR\Calendar\Decorator;
+
+/**
+ * Decorator to help with wrapping built children in another decorator
+ *
+ * @category  Date and Time
+ * @package   Calendar
+ * @author    Harry Fuecks <hfuecks@phppatterns.com>
+ * @author    Lorenzo Alberton <l.alberton@quipo.it>
+ * @copyright 2003-2007 Harry Fuecks, Lorenzo Alberton
+ * @license   http://www.debian.org/misc/bsd.license  BSD License (3 Clause)
+ * @link      http://pear.php.net/package/Calendar
+ * @access    public
+ */
+class Wrapper extends Decorator
+{
+    /**
+     * Constructs Calendar_Decorator_Wrapper
+     *
+     * @param object &$Calendar subclass of Calendar
+     *
+     * @access public
+     */
+    function __construct(&$Calendar)
+    {
+        parent::__construct($Calendar);
+    }
+
+    /**
+     * Wraps objects returned from fetch in the named Decorator class
+     *
+     * @param string $decorator name of Decorator class to wrap with
+     *
+     * @return object instance of named decorator
+     * @access public
+     */
+    function & fetch($decorator)
+    {
+        $Calendar = parent::fetch();
+        if ($Calendar) {
+            $ret = new $decorator($Calendar);
+        } else {
+            $ret = false;
+        }
+        return $ret;
+    }
+
+    /**
+     * Wraps the returned calendar objects from fetchAll in the named decorator
+     *
+     * @param string $decorator name of Decorator class to wrap with
+     *
+     * @return array
+     * @access public
+     */
+    function fetchAll($decorator)
+    {
+        $children = parent::fetchAll();
+        foreach ($children as $key => $Calendar) {
+            $children[$key] = new $decorator($Calendar);
+        }
+        return $children;
+    }
+}
