+++ src/Week.php
@@ -36,9 +36,9 @@
  * @version   CVS: $Id$
  * @link      http://pear.php.net/package/Calendar
  */
-namespace Pear\Calendar;
+namespace PEAR\Calendar;
 
-use Pear\Calendar\Table\Helper;
+use PEAR\Calendar\Table\Helper;
 
 /**
  * Represents a Week and builds Days in tabular format<br>
