+++ tests/HelperTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pear\Calendar\Test;
+namespace PEAR\Calendar\Test;
 
 use PHPUnit_Framework_TestCase;
